feat(clone-node, options): add ability to filter custom css properties from computed style
Recent Chromium revision added in custom css properties to the
computed style sheet which can lead to extremely slow
performance when computing sheets with hundreds of variables.
This uses (dummy) filtering to remove any property prefixed with '--'.
Also slightly changes the api of getStyleProperties to not double call getComputedStyle.

Related to #542, #536